Security patch

+++ src/main/java/com/adira/contact/service/ContactServiceImpl.java
@@ -3,15 +3,17 @@ package com.adira.contact.service;
 import com.adira.contact.common.Utils;
 import com.adira.contact.dto.ContactRequestDTO;
 import com.adira.contact.dto.ContactUpdateDTO;
+import com.adira.contact.entity.ApiResponse;
+import com.adira.contact.entity.Contact;
+import com.adira.contact.entity.User;
 import com.adira.contact.exception.CustomNotFoundException;
-import com.adira.contact.pojo.ApiResponse;
-import com.adira.contact.pojo.Contact;
-import com.adira.contact.pojo.User;
 import com.adira.contact.repository.ContactRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.BindingResult;
 
@@ -27,16 +29,20 @@ public class ContactServiceImpl implements ContactService {
     @Autowired
     private UserService userService;
 
-    private Utils utils = new Utils();
+    private final Utils utils = new Utils();
 
     @Override
     public List<Contact> getAllContacts() {
-        return contactRepository.findAll();
+        Long userId = getUserIdFromAuthentication();
+        return contactRepository.findByUserId(userId);
     }
 
     @Override
     public Contact getContactById(Long id) {
-        return contactRepository.findById(id).orElse(null);
+        Long userId = getUserIdFromAuthentication();
+        User user = getUserById(userId);
+        return contactRepository.findByIdAndUser(id, user)
+                .orElseThrow(() -> new CustomNotFoundException("Contact with id " + id + " not found"));
     }
 
     @Override
@@ -51,7 +57,8 @@ public class ContactServiceImpl implements ContactService {
             return utils.handleValidationErrors(bindingResult);
         }
 
-        Long userId = contactRequest.getUserId();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Long userId = Long.parseLong((String) authentication.getPrincipal());
 
         Optional<User> user = userService.getUserById(userId);
         if (!user.isPresent()) {
@@ -69,35 +76,25 @@ public class ContactServiceImpl implements ContactService {
     }
 
     @Override
-    public Contact updateContact(Long id, Contact contact) {
-        Optional<Contact> existingContactOptional = contactRepository.findById(id);
-
-        if (existingContactOptional.isPresent()) {
-            Contact existingContact = existingContactOptional.get();
-            existingContact.setAccountNumber(contact.getAccountNumber());
-            existingContact.setBankName(contact.getBankName());
-            existingContact.setContactName(contact.getContactName());
-            existingContact.setUser(contact.getUser());
-
-            return contactRepository.save(existingContact);
-        } else {
-            throw new CustomNotFoundException("Contact with id " + id + " not found");
-        }
-    }
-
-    @Override
     public ResponseEntity<?> deleteContact(Long id) {
-
         if (!doesContactExistById(id)) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND)
                     .body(new ApiResponse<>(404, "Contact Not Found", "API Contact Service", null));
+        }
+
+        Contact contact = contactRepository.findById(id)
+                .orElseThrow(() -> new CustomNotFoundException("Contact with id " + id + " not found"));
+        User user = getUserFromJWT();
+
+        if (contact.getUser().getId() != user.getId()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(new ApiResponse<>(403, "Contact not Belonging To User", "API Contact Service", null));
         }
 
         try {
             contactRepository.deleteById(id);
             return ResponseEntity.noContent().build();
         } catch (Exception e) {
-            // Handle any exceptions or errors during deletion
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .body(new ApiResponse<>(500, "Failed to delete contact", "API Contact Service", null));
         }
@@ -108,15 +105,37 @@ public class ContactServiceImpl implements ContactService {
         return contactRepository.existsById(id);
     }
 
-    public ResponseEntity<ApiResponse<Contact>> updateContactAggregate(Long id, ContactUpdateDTO contactUpdate) {
-        Optional<User> user = userService.getUserById(id);
+    private User getUserFromJWT() {
+        return getUserById(getUserIdFromAuthentication());
+    }
 
-        if (!user.isPresent()) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
-                    .body(new ApiResponse<>(400, "User not found", "API Contact Service", null));
+    private Long getUserIdFromAuthentication() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return Long.parseLong((String) authentication.getPrincipal());
+    }
+
+    private User getUserById(Long userId) {
+        return userService.getUserById(userId)
+                .orElseThrow(() -> new CustomNotFoundException("User with id " + userId + " not found"));
+    }
+
+    public ResponseEntity<ApiResponse<Contact>> updateContact(Long id, ContactUpdateDTO contactUpdate) {
+
+        Optional<Contact> contactOptional = contactRepository.findById(id);
+
+        if (!contactOptional.isPresent()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ApiResponse<>(404, "Contact Not Found", "API Contact Service", null));
         }
 
-        Contact contact = getContactById(id);
+        Contact contact = contactOptional.get();
+
+        User user = getUserFromJWT();
+
+        if (contact.getUser().getId() != user.getId()) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body(new ApiResponse<>(403, "Contact not Belonging To User", "API Contact Service", null));
+        }
 
         if (contactUpdate.getAccountNumber() != null) {
             contact.setAccountNumber(contactUpdate.getAccountNumber());
@@ -130,11 +149,27 @@ public class ContactServiceImpl implements ContactService {
             contact.setContactName(contactUpdate.getContactName());
         }
 
-        Contact updatedContact = updateContact(id, contact);
+        Contact updatedContact = contactRepository.save(contact);
+
         return updatedContact != null
                 ? ResponseEntity
                         .ok(new ApiResponse<>(200, "Contact " + id + " Updated", "API Contact Service", updatedContact))
                 : ResponseEntity.status(HttpStatus.BAD_REQUEST)
                         .body(new ApiResponse<>(400, "Failed", "API Contact Service", null));
+    }
+
+    @Override
+    public ResponseEntity<ApiResponse<List<Contact>>> findBySearchCriteria(String bankName, String accountNumber,
+            String contactName) {
+
+        List<Contact> contacts = contactRepository.findByCriteria(getUserFromJWT().getId(), bankName, accountNumber, contactName);
+
+        if (contacts.size() == 0) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body(new ApiResponse<>(404, "Contact Not Found", "API Contact Service", null));
+        }
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(new ApiResponse<>(200, "Contact Found", "API Contact Service", contacts));
     }
 }
